Detail and Lighting almost finished
So well, the map is almost finished with lighting and detail, altho more detail can be added later on, i warn you playtester map is 41 MB (i gotta see how to optimize this :c )

- There are black figures in the void area or the sky part... i've yet to fix this this is pending.

- Final area is not textured just lighted.

MOST of details and lights are final.

Still a bit to go for this map.

I might add a trigger_spawn below the yellow key for the 2 knights since spawning them earlier doesn't help much.